requirements.txt
blinker==1.6.2
click==8.1.3
exceptiongroup==1.1.1
Flask==2.3.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pydantic==1.10.8
pytest==7.3.1
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.6.2
Werkzeug==2.3.4
